Update PPN.cs
Degree and Name Validity Range fields are placed as PPN instead of ST as per the HL 2.7.1 spec causing a recursion in the constructor leading to a stack overflow using the public PPN(IMessage message, string description) constructor.  Changing them to be ST as per the HL 2.7.1 spec prevents the unintended recursion.